+++ packages/create/src/cli/prompts/promptForSchema.ts
