+++ packages/create/src/producers/producePreset.ts
@@ -3,7 +3,6 @@
 
 import { ProductionMode } from "../modes/types.js";
 import { AnyShape, InferredObject } from "../options.js";
-import { createSystemContext } from "../system/createSystemContext.js";
 import { createSystemContextWithAuth } from "../system/createSystemContextWithAuth.js";
 import { Creation } from "../types/creations.js";
 import { Preset } from "../types/presets.js";
